Refactor : 코드 가독성 위해 리팩토링을 진행한다.

+++ src/main/java/subway/dto/PathResponse.java
@@ -24,7 +24,7 @@ public class PathResponse {
         return stations;
     }
 
-    public double getDistance() {
+    public Long getDistance() {
         return distance;
     }
 }
